requirements.txt
PyAudio==0.2.14
websockets==12.0
janus==1.0.0
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
requests==2.32.3
httpx==0.27.0
fastapi
uvicorn[standard]
